+++ app/src/main/java/com/cmpt276/parentapp/model/TaskWithChild.java
@@ -12,5 +12,5 @@ public class TaskWithChild {
     @Embedded
     public Child child;
 
-    public int order;
+    public long order;
 }
